feature: scroll window into viewContact when itcha
